HOTFIX: Backport patch from KAFKA-6468
Read replication-offset-checkpoint once

Only read the high watermark checkpoint
file (replication-offset-checkpoint) once. Before this patch, this file
is read every time the broker handles LeaderAndIsrRequest. See
kafka.cluster.Partition#getOrCreateReplica(Int, Boolean).

On my local test cluster of three brokers with around 40k partitions,
the initial LeaderAndIsrRequest refers to every partition in the
cluster, and it can take 20 to 30 minutes to create all of the replicas
because the replication-offset-checkpoint is nearly 2MB.

Changing this code so that we only read this file once on startup
reduces the time to create all replicas to around one minute.

Conflicts:
  core/src/main/scala/kafka/cluster/Partition.scala
  core/src/main/scala/kafka/server/ReplicaManager.scala

RB=1217421
BUG=LIKAFKA-14552
G=Kafka-Code-Reviews
A=jqin