HOTFIX: cherry-pick fix for KAFKA-6469
KAFKA-6469 Batch ISR change notifications

We've failures in one of our test clusters as the partition count
started to climb north of 60k per broker. We had brokers writing child
nodes under /isr_change_notification that were larger than the
jute.maxbuffer size in ZooKeeper (1MB), causing the ZooKeeper server
to drop the controller's session, effectively bricking the cluster.

This can be mitigated by chunking ISR notifications to increase
the maximum number of partitions a broker can host, which is the
purpose of this patch.

ReplicaManager#maybePropagateIsrChanges() now batches the set of
TopicPartitions into sets of at most 3200 entries. This ensures that
the JSON payload is never more than around 912k at the worst case.

An integration test was added which propagates a set of 5000
TopicPartitions in two batches.

The dependency on System.getCurrentTimeMillis() was broken for testing
purposes.

Conflicts:
- core/src/main/scala/kafka/server/ReplicaManager.scala

RB=1217453
BUG=LIKAFKA-14552
G=Kafka-Code-Reviews
A=jqin