[LI-CHERRY-PICK] [ea4078e7] KAFKA-7311: Reset next batch expiry time on each poll loop TICKET = KAFKA-7311 LI_DESCRIPTION = preventative cherry-pick EXIT_CRITERIA = HASH [ea4078e72a737268cc45776a10d8c3a566be9b4a] ORIGINAL_DESCRIPTION =
Sender/RecordAccumulator never resets the next batch expiry time. Its always computed as the min of the current value and the expiry time for all batches being processed. This means that its always set to the expiry time of the first batch, and once that time has passed Sender starts spinning on epoll with a timeout of 0, which consumes a lot of CPU. This patch updates Sender to reset the next batch expiry time on each poll loop so that a new value reflecting the expiry time for the current set of batches is computed.

Reviewers: Guozhang Wang <wangguoz@gmail.com>
(cherry picked from commit ea4078e72a737268cc45776a10d8c3a566be9b4a)