[HOTFIX] LIKAFKA-19608 Refresh bootstrap server in ConsumerFetcherManager when TopicMetadata request fails
As part of GDPR the bootstrap server were fetched from the ZK for fetching TopicMetadata. But this
list is never refreshed. Recently all the broker in Kafka cluster were swapped out resulting
in TopicMetadata fetch request to fail as the requests were sent to old broker and never refreshed
after failed request.

RB=1464797
BUG=LIKAFKA-19608
R=zahuang
A=zahuang