[LI-HOTFIX] Add deletePartitionReassignment() in KafkaZkClient TICKET = LI_DESCRIPTION =
Upstream removes deletePartitionReassignment() in KafkaZkClient as a part of the fencing zk operations with controller epoch patch. However, cruise control depends on KafkaZkClient directly and relies on this API in some of its logics. This patch temporarily adds back the API in KafkaZkClient as a hotfix so that CC doesn’t break in compilation. 


RB=1537116
R=rrosenbl,agencer
A=rrosenbl

EXIT_CRITERIA = MANUAL ["After CC no longer relies on KafkaZkClient"]