cherry pick of: fc5f6b0e46ff81302b3e445fed0cdf454c942792 (#18)
* cherry pick of: fc5f6b0e46ff81302b3e445fed0cdf454c942792

MINOR: Add Timer to simplify timeout bookkeeping and use it in the consumer (#5087)

We currently do a lot of bookkeeping for timeouts which is both error-prone and distracting. This patch adds a new `Timer` class to simplify this logic and control unnecessary calls to system time. In particular, this helps with nested timeout operations. The consumer has been updated to use the new class.

Reviewers: Ismael Juma <ismael@juma.me.uk>, Guozhang Wang <wangguoz@gmail.com>

* Restore sensor functionality that was removed from patch.