[LI-CHERRY-PICK] [c29380f4] KAFKA-8715; Fix buggy reliance on state timestamp in static member.id generation (#7116) TICKET = KAFKA-8715 LI_DESCRIPTION = EXIT_CRITERIA = HASH [c29380f40d0db6d0c4a549117ba910f915a5053a] ORIGINAL_DESCRIPTION =
The bug is that we accidentally used the current state timestamp for the group instead of the real current time. When a group is first loaded, this timestamp is not initialized, so this resulted in a `NoSuchElementException`. Additionally this violated the intended uniqueness of the memberId, which could have broken the group instance fencing. Fix is made and unit test to make sure the timestamp is properly encoded within the returned member.id.

Reviewers: Ismael Juma <ismael@juma.me.uk>, Guozhang Wang <wangguoz@gmail.com>, Jason Gustafson <jason@confluent.io>
(cherry picked from commit c29380f40d0db6d0c4a549117ba910f915a5053a)