[LI-CHERRY-PICK] [24269269] KAFKA-7941: Catch TimeoutException in KafkaBasedLog worker thread (#6283) TICKET = KAFKA-7941 LI_DESCRIPTION = EXIT_CRITERIA = HASH [24269269b58cba9a826d0ab507e1fcf89f814666] ORIGINAL_DESCRIPTION =
When calling readLogToEnd(), the KafkaBasedLog worker thread should catch TimeoutException and log a warning, which can occur if brokers are unavailable, otherwise the worker thread terminates.

Includes an enhancement to MockConsumer that allows simulating exceptions not just when polling but also when querying for offsets, which is necessary for testing the fix.

Author: Paul Whalen <pgwhalen@gmail.com>
Reviewers: Randall Hauch <rhauch@gmail.com>, Arjun Satish <arjun@confluent.io>, Ryanne Dolan <ryannedolan@gmail.com>
(cherry picked from commit 24269269b58cba9a826d0ab507e1fcf89f814666)