[LI-CHERRY-PICK] [db8cb96a] KAFKA-8325; Remove batch from in-flight requests on MESSAGE_TOO_LARGE errors (#7176) TICKET = KAFKA-8325 LI_DESCRIPTION =
EXIT_CRITERIA = HASH [db8cb96ab151c52d41078032c94e7ff2091460a0]
ORIGINAL_DESCRIPTION =

This patch fixes a bug in the handling of MESSAGE_TOO_LARGE errors. The large batch is split, the smaller batches are re-added to the accumulator, and the batch is deallocated, but it was not removed from the list of in-flight batches. When the batch was eventually expired from the in-flight batches, the producer would try to deallocate it a second time, causing an error. This patch changes the behavior to correctly remove the batch from the list of in-flight requests.

Reviewers: Luke Stephenson, Ismael Juma <ismael@juma.me.uk>, Jason Gustafson <jason@confluent.io>
(cherry picked from commit db8cb96ab151c52d41078032c94e7ff2091460a0)